Adjust the structure for folders and files
