+++ src/UI/course/styles/courseCardStyle.scss
