+++ src/admin/adminUI/listModal/components/AdminStudentList.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Navigation from '../../../Navigation';
-import Footer from '../../../Footer';
+import Navigation from '../../../../UI/Navigation';
+import Footer from '../../../../UI/Footer';
 import StudentForm from './StudentForm';
 import DashMenu from '../../DashMenu';
 import PageNumber from './PageNumber';
