+++ src/admin/adminUI/listModal/components/CourseRow.js
