+++ src/admin/adminUI/listModal/components/PageNumber.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import NumberList from '../../../NumberList';
+import NumberList from '../../../../UI/NumberList';
 
 const numbers = ["<",1, 2, 3, 4, 5,">"];
 function PageNumber() {
