+++ src/admin/adminUI/searchModal/components/SearchBar.js
